Refactor function into various modules
